+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentOption+Images.swift
@@ -166,6 +166,10 @@ extension STPPaymentMethodType {
                     return .pm_type_blik
                 case .bacsDebit:
                     return .pm_type_us_bank
+                case .alipay:
+                    return .pm_type_alipay
+                case .OXXO:
+                    return .pm_type_oxxo
                 default:
                     return nil
                 }